combine animation state handling in hook
