+++ packages/dnb-eufemia/src/components/table/TableAccordionTr.tsx
@@ -1,8 +1,7 @@
-import React, { useCallback } from 'react'
+import React from 'react'
 import classnames from 'classnames'
-import { useHeightAnimation } from '../height-animation/useHeightAnimation'
-import { getClosestScrollViewElement } from '../../shared/component-helper'
-import { TableAccordionContext, TableContext } from './TableContext'
+import useTableAnimationHandler from './useTableAnimationHandler'
+import { TableContext } from './TableContext'
 import Td from './TableTd'
 
 export type TableAccordionTrProps = {
@@ -32,58 +31,21 @@ export default function TableAccordionTr(
   } = componentProps
 
   const allProps = React.useContext(TableContext)?.allProps
-  const tableAccordionContext = React.useContext(TableAccordionContext)
   const innerRef = React.useRef<HTMLDivElement>(null)
   const trRef = React.useRef<HTMLTableRowElement>(null)
-  const [ariaLive, setAriaLive] = React.useState(null)
 
-  const open = Boolean(expanded || tableAccordionContext?.trIsOpen)
-
-  const scrollViewHandler = useCallback((clip: boolean) => {
-    const scollView = getClosestScrollViewElement(
-      trRef.current
-    ) as HTMLElement
-    if (scollView instanceof HTMLElement) {
-      scollView.style.overflow = clip ? 'clip' : ''
-    }
-  }, [])
-
-  const onOpen = useCallback((state) => {
-    setAriaLive(state ? true : null)
-  }, [])
-
-  const onAnimationStart = useCallback(() => {
-    scrollViewHandler(true)
-  }, [scrollViewHandler])
-
-  const onAnimationEnd = useCallback(
-    (state) => {
-      const event = { target: trRef.current }
-      switch (state) {
-        case 'opened':
-          tableAccordionContext.onOpened?.(event)
-          break
-
-        case 'closed':
-          tableAccordionContext.onClosed?.(event)
-          break
-      }
-
-      scrollViewHandler(false)
-    },
-    [scrollViewHandler, tableAccordionContext]
-  )
-
-  const { isInDOM, isAnimating, isVisibleParallax, firstPaintStyle } =
-    useHeightAnimation(innerRef, {
-      open,
-      animate: Boolean(
-        !noAnimation && !tableAccordionContext?.noAnimation
-      ),
-      onOpen,
-      onAnimationStart,
-      onAnimationEnd,
-    })
+  const {
+    ariaLive,
+    isInDOM,
+    isAnimating,
+    isVisibleParallax,
+    firstPaintStyle,
+  } = useTableAnimationHandler({
+    innerRef,
+    trRef,
+    expanded,
+    noAnimation,
+  })
 
   const expandColumn = (
     <Td>
